# packages/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "内部 Connector 服务：catalog、配置管理、OAuth 连接与聚合 MCP。\n管理端 /admin 使用 HttpOnly SameSite=Strict Cookie；CONNECT_IT_BASE_URL 为 HTTPS 时固定设置 Secure。浏览器客户端须启用 credentials=include，且写请求须携带 X-Connect-It-CSRF: 1。Swagger 2 无法形式化表达 Cookie 鉴权。",
        "title": "connect-it API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/api-tokens": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "列出 API token（不含明文）",
                "operationId": "listApiTokens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/authsvc.APITokenView"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "创建 API token（明文仅在响应中出现一次）",
                "operationId": "createApiToken",
                "parameters": [
                    {
                        "type": "string",
                        "description": "管理台同源写请求固定值 1",
                        "name": "X-Connect-It-CSRF",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "token 名称",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.createTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/api-tokens/{id}": {
            "delete": {
                "tags": [
                    "admin"
                ],
                "summary": "撤销 API token",
                "operationId": "deleteApiToken",
                "parameters": [
                    {
                        "type": "string",
                        "description": "管理台同源写请求固定值 1",
                        "name": "X-Connect-It-CSRF",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token id（uuid）",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/connectors/{type}/mcp:verify": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "实测 Connector 的全部 Remote MCP server 并比对 tool 映射",
                "operationId": "verifyMcp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "管理台同源写请求固定值 1",
                        "name": "X-Connect-It-CSRF",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "connector_type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/mcp-sessions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mcp"
                ],
                "summary": "签发短期 MCP session token（固化 alias、Tool grant 与授权代际）",
                "operationId": "createMcpSession",
                "parameters": [
                    {
                        "description": "连接绑定与不可变 grant 快照；省略 allowlist 默认当前 read，[] 为零 Tool，null 拒绝；write/destructive 必须显式列出；ttl_seconds 默认 3600、上限 86400",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createMCPSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.createMCPSessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/oauth/callback": {
            "get": {
                "tags": [
                    "connections"
                ],
                "summary": "OAuth 回调（provider 跳转回来，无鉴权，靠一次性 state）",
                "operationId": "oauthCallback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "授权发起时生成的 state",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "授权码",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "provider 返回的错误码",
                        "name": "error",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error 是稳定、机器可读的错误码。",
                    "type": "string"
                },
                "message": {
                    "description": "Message 可安全返回调用方，不含 Provider 原始响应或 credential 值。",
                    "type": "string"
                },
                "retry_after_seconds": {
                    "description": "RetryAfterSeconds 仅在已知正数重试延迟时出现。",
                    "type": "integer"
                },
                "temporary": {
                    "description": "Temporary 表示稍后重试是否可能成功。",
                    "type": "boolean"
                }
            }
        },
...
        "api.createMCPSessionRequest": {
            "type": "object",
            "properties": {
                "connections": {
                    "description": "Connections 将 Session-local alias 映射到持久 Connection UUID。",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "tool_allowlist": {
                    "description": "ToolAllowlist 区分三态：省略时固化当前 read Tool，[] 授权零 Tool，null\n非法；write/destructive Tool 必须显式列出。",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ttl_seconds": {
                    "description": "TTLSeconds 是 Session 有效期；0 使用默认 3600 秒，上限 86400 秒。",
                    "type": "integer",
                    "maximum": 86400,
                    "minimum": 0
                }
            }
        },
        "api.createMCPSessionResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "api.createTokenRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "api.createTokenResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "authsvc.APITokenView": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "revoked_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}